refactor track fitting algorithm
