+++ Examples/Algorithms/TrackFitting/include/ActsExamples/TrackFitting/KalmanFitterFunction.hpp
@@ -1,0 +1,24 @@
+// This file is part of the Acts project.
+//
+// Copyright (C) 2019-2021 CERN for the benefit of the Acts project
+//
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+#pragma once
+
+#include "ActsExamples/TrackFitting/TrackFittingAlgorithm.hpp"
+
+namespace ActsExamples {
+
+std::shared_ptr<TrackFittingAlgorithm::TrackFitterFunction>
+makeKalmanFitterFunction(
+    std::shared_ptr<const Acts::TrackingGeometry> trackingGeometry,
+    std::shared_ptr<const Acts::MagneticFieldProvider> magneticField,
+    bool multipleScattering = true, bool energyLoss = true,
+    double reverseFilteringMomThreshold = 0.0,
+    Acts::FreeToBoundCorrection freeToBoundCorrection =
+        Acts::FreeToBoundCorrection());
+
+}  // namespace ActsExamples
