+++ Examples/Algorithms/TrackFitting/include/ActsExamples/TrackFitting/TrackFittingAlgorithm.hpp
@@ -59,20 +59,7 @@ class TrackFittingAlgorithm final : public BareAlgorithm {
     virtual TrackFitterResult operator()(
         const std::vector<std::reference_wrapper<const IndexSourceLink>>&,
         const TrackParameters&, const GeneralFitterOptions&,
-        std::shared_ptr<Acts::VectorMultiTrajectory>& trajectory) const = 0;
-  };
-
-  /// Fit function that takes the above parameters plus a sorted surface
-  /// sequence for the DirectNavigator to follow
-  /// @note This is separated into a virtual interface to keep compilation units
-  /// small
-  class DirectedTrackFitterFunction {
-   public:
-    virtual ~DirectedTrackFitterFunction() = default;
-    virtual TrackFitterResult operator()(
-        const std::vector<std::reference_wrapper<const IndexSourceLink>>&,
-        const TrackParameters&, const GeneralFitterOptions&,
-        const std::vector<const Acts::Surface*>&,
+        const std::optional<std::vector<const Acts::Surface*>>&,
         std::shared_ptr<Acts::VectorMultiTrajectory>& trajectory) const = 0;
   };
 
@@ -91,8 +78,6 @@ class TrackFittingAlgorithm final : public BareAlgorithm {
     std::string outputTrajectories;
     /// Type erased fitter function.
     std::shared_ptr<TrackFitterFunction> fit;
-    /// Type erased direct navigation fitter function
-    std::shared_ptr<DirectedTrackFitterFunction> dFit;
     /// Tracking geometry for surface lookup
     std::shared_ptr<const Acts::TrackingGeometry> trackingGeometry;
     /// Pick a single track for debugging (-1 process all tracks)
@@ -114,68 +99,18 @@ class TrackFittingAlgorithm final : public BareAlgorithm {
   /// Get readonly access to the config parameters
   const Config& config() const { return m_cfg; }
 
-  /// Create the track fitter function implementation.
-  ///
-  /// The magnetic field is intentionally given by-value since the variant
-  /// contains shared_ptr anyways.
-  /// @param trackingGeometry
-  /// @param multipleScattering correct for MCS (mainly for debugging)
-  /// @param energyLoss correct for e-loss
-  /// @param reverseFilteringMomThreshold at which threshold
-  /// @param freeToBoundCorrection Correction for non-linearity effect during transform from free to bound
-  static std::shared_ptr<TrackFitterFunction> makeKalmanFitterFunction(
-      std::shared_ptr<const Acts::TrackingGeometry> trackingGeometry,
-      std::shared_ptr<const Acts::MagneticFieldProvider> magneticField,
-      bool multipleScattering = true, bool energyLoss = true,
-      double reverseFilteringMomThreshold = 0.0,
-      Acts::FreeToBoundCorrection freeToBoundCorrection =
-          Acts::FreeToBoundCorrection());
-
-  static std::shared_ptr<DirectedTrackFitterFunction> makeKalmanFitterFunction(
-      std::shared_ptr<const Acts::MagneticFieldProvider> magneticField,
-      bool multipleScattering = true, bool energyLoss = true,
-      double reverseFilteringMomThreshold = 0.0,
-      Acts::FreeToBoundCorrection freeToBoundCorrection =
-          Acts::FreeToBoundCorrection());
-
-  static std::shared_ptr<TrackFitterFunction> makeGsfFitterFunction(
-      std::shared_ptr<const Acts::TrackingGeometry> trackingGeometry,
-      std::shared_ptr<const Acts::MagneticFieldProvider> magneticField,
-      std::size_t maxComponents = 4, bool abortOnError = true,
-      bool disableAllMaterialHandling = false);
-
-  static std::shared_ptr<DirectedTrackFitterFunction> makeGsfFitterFunction(
-      std::shared_ptr<const Acts::MagneticFieldProvider> magneticField,
-      std::size_t maxComponents = 4, bool abortOnError = true,
-      bool disableAllMaterialHandling = false);
+  // /// Create the track fitter function implementation.
+  // ///
+  // /// The magnetic field is intentionally given by-value since the variant
+  // /// contains shared_ptr anyways.
+  // /// @param trackingGeometry
+  // /// @param multipleScattering correct for MCS (mainly for debugging)
+  // /// @param energyLoss correct for e-loss
+  // /// @param reverseFilteringMomThreshold at which threshold
+  // /// @param freeToBoundCorrection Correction for non-linearity effect during transform from free to bound
 
  private:
-  /// Helper function to call correct FitterFunction
-  TrackFitterResult fitTrack(
-      const std::vector<std::reference_wrapper<
-          const ActsExamples::IndexSourceLink>>& sourceLinks,
-      const ActsExamples::TrackParameters& initialParameters,
-      const GeneralFitterOptions& options,
-      const std::vector<const Acts::Surface*>& surfSequence,
-      std::shared_ptr<Acts::VectorMultiTrajectory>& trajectory) const;
-
   Config m_cfg;
 };
-
-inline ActsExamples::TrackFittingAlgorithm::TrackFitterResult
-ActsExamples::TrackFittingAlgorithm::fitTrack(
-    const std::vector<std::reference_wrapper<
-        const ActsExamples::IndexSourceLink>>& sourceLinks,
-    const ActsExamples::TrackParameters& initialParameters,
-    const ActsExamples::TrackFittingAlgorithm::GeneralFitterOptions& options,
-    const std::vector<const Acts::Surface*>& surfSequence,
-    std::shared_ptr<Acts::VectorMultiTrajectory>& trajectory) const {
-  if (m_cfg.directNavigation) {
-    return (*m_cfg.dFit)(sourceLinks, initialParameters, options, surfSequence,
-                         trajectory);
-  }
-
-  return (*m_cfg.fit)(sourceLinks, initialParameters, options, trajectory);
-}
 
 }  // namespace ActsExamples
