+++ Examples/Algorithms/TrackFitting/src/TrackFittingAlgorithm.cpp
@@ -112,6 +112,8 @@ ActsExamples::ProcessCode ActsExamples::TrackFittingAlgorithm::execute(
     // Clear & reserve the right size
     trackSourceLinks.clear();
     trackSourceLinks.reserve(protoTrack.size());
+    surfSequence.clear();
+    surfSequence.reserve(protoTrack.size());
 
     // Fill the source links via their indices from the container
     for (auto hitIndex : protoTrack) {
@@ -127,9 +129,12 @@ ActsExamples::ProcessCode ActsExamples::TrackFittingAlgorithm::execute(
       }
     }
 
-    ACTS_DEBUG("Invoke fitter");
-    auto result =
-        fitTrack(trackSourceLinks, initialParams, options, surfSequence, mtj);
+    ACTS_DEBUG("Invoke direct fitter for track " << itrack);
+    auto result = m_cfg.directNavigation
+                      ? (*m_cfg.fit)(trackSourceLinks, initialParams, options,
+                                     surfSequence, mtj)
+                      : (*m_cfg.fit)(trackSourceLinks, initialParams, options,
+                                     std::nullopt, mtj);
 
     if (result.ok()) {
       // Get the fit output object
