+++ Examples/Python/src/TrackFitting.cpp
@@ -8,6 +8,8 @@
 
 #include "Acts/MagneticField/MagneticFieldProvider.hpp"
 #include "Acts/Plugins/Python/Utilities.hpp"
+#include "ActsExamples/TrackFitting/GsfFitterFunction.hpp"
+#include "ActsExamples/TrackFitting/KalmanFitterFunction.hpp"
 #include "ActsExamples/TrackFitting/SurfaceSortingAlgorithm.hpp"
 #include "ActsExamples/TrackFitting/TrackFittingAlgorithm.hpp"
 
@@ -35,56 +37,15 @@ void addTrackFitting(Context& ctx) {
     using Alg = ActsExamples::TrackFittingAlgorithm;
     using Config = Alg::Config;
 
-    auto alg =
-        py::class_<Alg, BareAlgorithm, std::shared_ptr<Alg>>(
-            mex, "TrackFittingAlgorithm")
-            .def(py::init<const Alg::Config&, Acts::Logging::Level>(),
-                 py::arg("config"), py::arg("level"))
-            .def_property_readonly("config", &Alg::config)
-            .def_static("makeKalmanFitterFunction",
-                        py::overload_cast<
-                            std::shared_ptr<const Acts::TrackingGeometry>,
-                            std::shared_ptr<const Acts::MagneticFieldProvider>,
-                            bool, bool, double, Acts::FreeToBoundCorrection>(
-                            &Alg::makeKalmanFitterFunction),
-                        py::arg("trackingGeometry"), py::arg("magneticField"),
-                        py::arg("multipleScattering"), py::arg("energyLoss"),
-                        py::arg("reverseFilteringMomThreshold"),
-                        py::arg("freeToBoundCorrection"))
-            .def_static("makeKalmanFitterFunction",
-                        py::overload_cast<
-                            std::shared_ptr<const Acts::MagneticFieldProvider>,
-                            bool, bool, double, Acts::FreeToBoundCorrection>(
-                            &Alg::makeKalmanFitterFunction),
-                        py::arg("magneticField"), py::arg("multipleScattering"),
-                        py::arg("energyLoss"),
-                        py::arg("reverseFilteringMomThreshold"),
-                        py::arg("freeToBoundCorrection"))
-            .def_static(
-                "makeGsfFitterFunction",
-                py::overload_cast<
-                    std::shared_ptr<const Acts::TrackingGeometry>,
-                    std::shared_ptr<const Acts::MagneticFieldProvider>,
-                    std::size_t, bool, bool>(&Alg::makeGsfFitterFunction),
-                py::arg("trackingGeometry"), py::arg("magneticField"),
-                py::arg("maxComponents"), py::arg("abortOnError"),
-                py::arg("disableAllMaterialHandling"))
-            .def_static(
-                "makeGsfFitterFunction",
-                py::overload_cast<
-                    std::shared_ptr<const Acts::MagneticFieldProvider>,
-                    std::size_t, bool, bool>(&Alg::makeGsfFitterFunction),
-                py::arg("magneticField"), py::arg("maxComponents"),
-                py::arg("abortOnError"), py::arg("disableAllMaterialHandling"));
+    auto alg = py::class_<Alg, BareAlgorithm, std::shared_ptr<Alg>>(
+                   mex, "TrackFittingAlgorithm")
+                   .def(py::init<const Alg::Config&, Acts::Logging::Level>(),
+                        py::arg("config"), py::arg("level"))
+                   .def_property_readonly("config", &Alg::config);
 
     py::class_<TrackFittingAlgorithm::TrackFitterFunction,
                std::shared_ptr<TrackFittingAlgorithm::TrackFitterFunction>>(
         alg, "TrackFitterFunction");
-
-    py::class_<
-        TrackFittingAlgorithm::DirectedTrackFitterFunction,
-        std::shared_ptr<TrackFittingAlgorithm::DirectedTrackFitterFunction>>(
-        alg, "DirectedTrackFitterFunction");
 
     auto c = py::class_<Config>(alg, "Config").def(py::init<>());
 
@@ -96,10 +57,30 @@ void addTrackFitting(Context& ctx) {
     ACTS_PYTHON_MEMBER(inputInitialTrackParameters);
     ACTS_PYTHON_MEMBER(outputTrajectories);
     ACTS_PYTHON_MEMBER(fit);
-    ACTS_PYTHON_MEMBER(dFit);
     ACTS_PYTHON_MEMBER(trackingGeometry);
     ACTS_PYTHON_MEMBER(pickTrack);
     ACTS_PYTHON_STRUCT_END();
+
+    mex.def(
+        "makeKalmanFitterFunction",
+        py::overload_cast<std::shared_ptr<const Acts::TrackingGeometry>,
+                          std::shared_ptr<const Acts::MagneticFieldProvider>,
+                          bool, bool, double, Acts::FreeToBoundCorrection>(
+            &ActsExamples::makeKalmanFitterFunction),
+        py::arg("trackingGeometry"), py::arg("magneticField"),
+        py::arg("multipleScattering"), py::arg("energyLoss"),
+        py::arg("reverseFilteringMomThreshold"),
+        py::arg("freeToBoundCorrection"));
+
+    mex.def(
+        "makeGsfFitterFunction",
+        py::overload_cast<std::shared_ptr<const Acts::TrackingGeometry>,
+                          std::shared_ptr<const Acts::MagneticFieldProvider>,
+                          std::size_t, bool, bool>(
+            &ActsExamples::makeGsfFitterFunction),
+        py::arg("trackingGeometry"), py::arg("magneticField"),
+        py::arg("maxComponents"), py::arg("abortOnError"),
+        py::arg("disableAllMaterialHandling"));
   }
 
   {
