+++ Examples/Run/Reconstruction/Common/RecTruthTracks.cpp
@@ -22,6 +22,7 @@
 #include "ActsExamples/MagneticField/MagneticFieldOptions.hpp"
 #include "ActsExamples/Options/CommonOptions.hpp"
 #include "ActsExamples/Reconstruction/ReconstructionBase.hpp"
+#include "ActsExamples/TrackFitting/KalmanFitterFunction.hpp"
 #include "ActsExamples/TrackFitting/SurfaceSortingAlgorithm.hpp"
 #include "ActsExamples/TrackFitting/TrackFittingAlgorithm.hpp"
 #include "ActsExamples/TrackFitting/TrackFittingOptions.hpp"
@@ -146,12 +147,7 @@ int runRecTruthTracks(int argc, char* argv[],
   fitter.directNavigation = dirNav;
   fitter.pickTrack = vm["fit-pick-track"].as<int>();
   fitter.trackingGeometry = trackingGeometry;
-  fitter.dFit = TrackFittingAlgorithm::makeKalmanFitterFunction(
-      magneticField, vm["fit-multiple-scattering-correction"].as<bool>(),
-      vm["fit-energy-loss-correction"].as<bool>(), reverseFilteringMomThreshold,
-      Acts::FreeToBoundCorrection(
-          vm["fit-ftob-nonlinear-correction"].as<bool>()));
-  fitter.fit = TrackFittingAlgorithm::makeKalmanFitterFunction(
+  fitter.fit = makeKalmanFitterFunction(
       trackingGeometry, magneticField,
       vm["fit-multiple-scattering-correction"].as<bool>(),
       vm["fit-energy-loss-correction"].as<bool>(), reverseFilteringMomThreshold,
